Move confirm_rebase into core
