+++ src/core/modules/confirm_rebase/mod.rs
