+++ src/core/modules/mod.rs
@@ -1,3 +1,4 @@
 mod confirm_abort;
+mod confirm_rebase;
 
-pub use self::confirm_abort::ConfirmAbort;
+pub use self::{confirm_abort::ConfirmAbort, confirm_rebase::ConfirmRebase};
